thirdeye/requirements.txt
django-cors-headers
channels
setuptools
djangorestframework_simplejwt
drf-yasg
django
djangorestframework
google-auth
requests
opencv-python
face_recognition
mtcnn
keras
tensorflow
numpy
python-decouple


